clk: Port Linux common clock framework [CCF] for imx6q to U-boot (tag: 5.0-rc3)
This commit brings the files from Linux kernel to provide clocks support
as it is used on the Linux kernel with common clock framework [CCF] setup.

The directory structure has been preserved. The ported code only supports
reading information from PLL, MUX, Divider, etc and enabling/disabling
the clocks USDHCx/ECSPIx depending on used bus. Moreover, it is agnostic
to the alias numbering as the information about the clock is read from
device tree.

One needs to pay attention to the comments indicating necessary for U-boot's
DM changes.

If needed the code can be extended to support the "set" part of the clock
management.


Series-to: sbabic, Fabio Estevam <festevam@gmail.com>, marex, sjg, trini, u-boot, jagan, Peng Fan <peng.fan@nxp.com>, Marcel Ziswiler <marcel.ziswiler@toradex.com>, Adam Ford <aford173@gmail.com>
Series-version: 3
Cover-letter:
clk: Port Linux common clock framework [CCF] to U-boot (tag: 5.0-rc3)
This patch series brings the files from Linux kernel to provide clocks support
as it is used on the Linux kernel with common clock framework [CCF] setup.

This series also fixes several problems with current clocks and provides
sandbox tests for functions addded to clk-uclass.c file.

Design decisions/issues:
=========================

- U-boot's DM for clk differs from Linux CCF. The most notably difference
is the lack of support for hierarchical clocks and "clock as a manager
driver" (single clock DTS node acts as a starting point for all other
clocks).

- The clk_get_rate() now caches the previously read data (no need for
recursive access.

- On purpose the "manager" clk driver (clk-imx6q.c) is not using large
table to store pointers to clocks - e.g. clk[IMX6QDL_CLK_USDHC2_SEL] = ....
Instead we use udevice's linked list for the same class (UCLASS_CLK).
The rationale - when porting the code as is from Linux, one would need
~1KiB of RAM to store it. This is way too much if we do plan to use this
driver in SPL.

- The "central" structure of this patch series is struct udevice and its
driver_data field contains the struct clk pointer (to the originally created
one).

- Up till now U-boot's driver model's CLK operates on udevice (main access to
  clock is by udevice ops)
  In the CCF the access to struct clk (comprising pointer to *dev) is
  possible via dev_get_driver_data()

  Storing back pointer (from udevice to struct clk) as driver_data is a
  convention for CCF.

- I could use *private_alloc_size to allocate driver's 'private"
  structures (dev->priv) for e.g. divider (struct clk_divider *divider)
  for IMX6Q clock, but this would change the original structure of the CCF code.

The question is if it would be better to use private_alloc_size (and
dev->private) or stay with driver_data.
The former requires some rewritting in CCF original code (to remove
(c)malloc, etc), but comply with u-boot DM. The latter allows re-using the
CCF code as is, but introduces some convention special for CCF (I'm not
sure thought if dev->priv is NOT another convention as well).

- I've added the clk_get_parent(), which reads parent's dev->driver_data to
  provide parent's struct clk pointer. This seems the easiest way to get
  child/parent relationship for struct clk in U-boot's udevice
  based clocks.

- For tests I had to "emulate" CCF code structure to test functionality of clk_get_parent_rate()
  and clk_get_by_id(). Those functions will not work properly with
  "standard" (i.e. non CCF) clock setup(with not set dev->driver_data to
  struct clk).

END

Signed-off-by: Lukasz Majewski <lukma@denx.de>